requirements.txt
# C&D Waste Classifier — Python Dependencies
# Install with:  pip install -r requirements.txt

torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
